Feat: Overall Structural Changes
