+++ src/components/MessagePage/Profile.module.css
@@ -34,16 +34,23 @@
 
 .delete {
   position: absolute;
-  z-index: 5;
-  width: 6rem;
-}
-
-.delete:hover {
-  opacity: 0.5;
+  width: 8rem;
+  height: 8rem;
+  padding-top: 3rem;
+  border-radius: 50%;
+  background-color: var(--white);
+  color: var(--black);
+  font-weight: bold;
+  opacity: 0.7;
+  text-align: center;
 }
 
 .customImgContainer {
+  position: relative;
   display: flex;
-  align-items: center;
-  justify-content: center;
+  width: 8rem;
+  height: 8rem;
+
+  /* align-items: center;
+  justify-content: center; */
 }
